add remove user from days button
